[refactor] 파일 경로 수정 및 대문자 수정 close #51
