+++ pages/channels.tsx
@@ -2,7 +2,7 @@ import useTranslation from 'next-translate/useTranslation';
 
 import PageHeader from 'components/global/PageHeader';
 
-import styles from 'styles/pages/channels.module.scss';
+import styles from 'styles/channels/Channels.module.scss';
 
 export default function channels() {
   const { t } = useTranslation('channels');
