+++ pages/friends.tsx
@@ -2,7 +2,7 @@ import useTranslation from 'next-translate/useTranslation';
 
 import PageHeader from 'components/global/PageHeader';
 
-import styles from 'styles/pages/friends.module.scss';
+import styles from 'styles/friends/Friends.module.scss';
 
 export default function friends() {
   const { t } = useTranslation('friends');
