+++ pages/leaderboard.tsx
@@ -2,7 +2,7 @@ import useTranslation from 'next-translate/useTranslation';
 
 import PageHeader from 'components/global/PageHeader';
 
-import styles from 'styles/pages/leaderboard.module.scss';
+import styles from 'styles/leaderboard/Leaderboard.module.scss';
 
 export default function leaderboard() {
   const { t } = useTranslation('leaderboard');
