+++ pages/matchHistory.tsx
@@ -2,7 +2,7 @@ import useTranslation from 'next-translate/useTranslation';
 
 import PageHeader from 'components/global/PageHeader';
 
-import styles from 'styles/pages/matchHistory.module.scss';
+import styles from 'styles/matchHistory/MatchHistory.module.scss';
 
 export default function matchHistory() {
   const { t } = useTranslation('matchHistory');
