+++ pages/myPage.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 import PageHeader from 'components/global/PageHeader';
 import MyPageFrame from 'components/myPage/MyPageFrame';
 
-import styles from 'styles/pages/myPage.module.scss';
+import styles from 'styles/myPage/MyPage.module.scss';
 
 export default function MyPage() {
   const { t } = useTranslation(['page']);
