+++ pages/settings.tsx
@@ -3,7 +3,7 @@ import useTranslation from 'next-translate/useTranslation';
 import PageHeader from 'components/global/PageHeader';
 import SettingsFrame from 'components/settings/SettingsFrame';
 
-import styles from 'styles/settings/settings.module.scss';
+import styles from 'styles/settings/Settings.module.scss';
 
 export default function settings() {
   const { t } = useTranslation('settings');
